phy: hisilicon: remove bus logic, use parent's regmap directly
We are moving bus write/read logic to a dedicated bus driver. So we
should get regmap from the parent device and use it to perform register
write operation. PHY_TYPE enums are not needed anymore, this driver is
generic now.

If it fails to retrive regmap from parent, it means the PHY configuration
interface are not attached to the PERICTRL USB2PHY test bus. Fallback to
 simple MMIO under such cases.

Signed-off-by: Yang Xiwen <forbidden405@outlook.com>