FROMLIST: dmaengine: qcom: bam_dma: Avoid accessing BAM_REVISION on remote BAM
Commit '57a7138d0627 ("dmaengine: qcom: bam_dma: Avoid writing
unavailable register")' made this read unconditional, in order to
identify if the instance is BAM-NDP or BAM-Lite.
But the BAM_REVISION register is not accessible on remotely managed BAM
instances and attempts to access it causes the system to crash.

Move the access back to be conditional and expand the checks that was
introduced to restore the old behavior when no revision information is
available.

Fixes: 57a7138d0627 ("dmaengine: qcom: bam_dma: Avoid writing unavailable register")
Reported-by: Georgi Djakov <djakov@kernel.org>
Closes: https://lore.kernel.org/lkml/9ef3daa8-cdb1-49f2-8d19-a72d6210ff3a@kernel.org/
Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>
Tested-by: Sam Day <me@samcday.com>
Tested-by: Georgi Djakov <djakov@kernel.org> # db845c
Tested-by: David Heidelberg <david@ixit.cz> # OnePlus 6T
Link: https://lore.kernel.org/r/20250207-bam-read-fix-v1-1-027975cf1a04@oss.qualcomm.com